// lib/relay/__generated__/postsScreenFollowingFeedQuery.graphql.ts
export type postsScreenFollowingFeedQuery$variables = Record<PropertyKey, never>;
export type postsScreenFollowingFeedQuery$data = {
  readonly user: {
    readonly followingCount: number;
    readonly id: string;
  } | null | undefined;
  readonly " $fragmentSpreads": FragmentRefs<"postsScreenFollowingFeedFragment">;
};
export type postsScreenFollowingFeedQuery = {
  response: postsScreenFollowingFeedQuery$data;
  variables: postsScreenFollowingFeedQuery$variables;
};